+++ src/pages/Main/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import * as s from "./styles";
-import LayoutTemplate from '../../components/LayoutTemplate';
-import GoogleLoginButton from '../../components/GoogleLoginButton';
-import KaKaoLogin from '../../components/KaKaoLogin';
+import LayoutTemplate from '../../components/Common/LayoutTemplate';
+import GoogleLoginButton from '../../components/Login/GoogleLoginButton';
+import KaKaoLogin from '../../components/Login/KaKaoLogin';
 import MainTab from '../../components/Main/MainTab';
 import TabImgs from '../../components/Main/TabImgs';
 const MainPage = () => {
